// frontend/src/api-client/model/lint-request-dto.ts
/**
 * Inkweld API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LintRequestDto { 
    /**
     * The paragraph text to be checked for errors
     */
    paragraph: string;
    /**
     * The writing style to check against
     */
    style: string;
    /**
     * The level of linting strictness
     */
    level: LintRequestDto.LevelEnum;
}
export namespace LintRequestDto {
    export type LevelEnum = 'low' | 'medium' | 'high';
    export const LevelEnum = {
        Low: 'low' as LevelEnum,
        Medium: 'medium' as LevelEnum,
        High: 'high' as LevelEnum
    };
}