// frontend/src/api-client/model/lint-response-dto.ts
/**
 * Inkweld API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StyleRecommendationDto } from './style-recommendation-dto';
import { CorrectionDto } from './correction-dto';


export interface LintResponseDto { 
    /**
     * The original paragraph that was checked
     */
    original_paragraph: string;
    /**
     * List of text corrections with position information
     */
    corrections: Array<CorrectionDto>;
    /**
     * List of style recommendations for the paragraph
     */
    style_recommendations: Array<StyleRecommendationDto>;
    /**
     * The source of the linting service (openai or languagetool)
     */
    source: LintResponseDto.SourceEnum;
}
export namespace LintResponseDto {
    export type SourceEnum = 'openai' | 'languagetool';
    export const SourceEnum = {
        Openai: 'openai' as SourceEnum,
        Languagetool: 'languagetool' as SourceEnum
    };
}